Mini-Batch k-means for large single cell sequencing data

This is an implementation of the mini-batch k-means algorithm of Sculley (2010) for large single cell sequencing data with the dimensionality reduction results as input in the reducedDim() slot.

Usage

mbkmeans(x, ...)

# S4 method for SummarizedExperiment
mbkmeans(x, whichAssay = 1, ...)

# S4 method for SingleCellExperiment
mbkmeans(x, reduceMethod = "PCA", whichAssay = 1, ...)

# S4 method for LinearEmbeddingMatrix
mbkmeans(x, ...)

# S4 method for ANY
mbkmeans(
  x,
  clusters,
  batch_size = min(500, NCOL(x)),
  max_iters = 100,
  num_init = 1,
  init_fraction = batch_size/NCOL(x),
  initializer = "kmeans++",
  compute_labels = TRUE,
  calc_wcss = FALSE,
  early_stop_iter = 10,
  verbose = FALSE,
  CENTROIDS = NULL,
  tol = 1e-04,
  BPPARAM = BiocParallel::SerialParam(),
  ...
)

Arguments

x

The object on which to run mini-batch k-means. It can be a matrix-like object (e.g., matrix, Matrix, DelayedMatrix, HDF5Matrix) with genes in the rows and samples in the columns. Specialized methods are defined for SummarizedExperiment and SingleCellExperiment.

...

passed to `blockApply`.

whichAssay

The assay to use as input to mini-batch k-means. If x is a SingleCellExperiment, this is ignored unless reduceMethod = NA.

reduceMethod

Name of dimensionality reduction results to use as input to mini-batch k-means. Set to NA to use the full matrix.

clusters

the number of clusters

batch_size

the size of the mini batches. By default, it equals the minimum between the number of observations and 500.

max_iters

the maximum number of clustering iterations

num_init

number of times the algorithm will be run with different centroid seeds

init_fraction

proportion of data to use for the initialization centroids (applies if initializer is kmeans++ ). Should be a float number between 0.0 and 1.0. By default, it uses the relative batch size.

initializer

the method of initialization. One of kmeans++ and random. See details for more information

compute_labels

logcical indicating whether to compute the final cluster labels.

calc_wcss

logical indicating whether the per-cluster WCSS is computed. Ignored if `compute_labels = FALSE`.

early_stop_iter

continue that many iterations after calculation of the best within-cluster-sum-of-squared-error

verbose

either TRUE or FALSE, indicating whether progress is printed during clustering

CENTROIDS

a matrix of initial cluster centroids. The rows of the CENTROIDS matrix should be equal to the number of clusters and the columns should be equal to the columns of the data

tol

a float number. If, in case of an iteration (iteration > 1 and iteration < max_iters) 'tol' is greater than the squared norm of the centroids, then kmeans has converged

BPPARAM

See the `BiocParallel` package. Only the label assignment is done in parallel.

Value

A list with the following attributes: centroids, WCSS_per_cluster, best_initialization, iters_per_initialization.

a list with the following attributes: centroids, WCSS_per_cluster, best_initialization, iters_per_initialization

Details

The implementation is largely based on the MiniBatchKmeans function of the ClusterR package. The contribution of this package is to provide support for on-disk data representations such as HDF5, through the use of DelayedMatrix and HDF5Matrix objects, as well as for sparse data representation through the classes of the Matrix package. We also provide high-level methods for objects of class SummarizedExperiment, SingleCellExperiment, and LinearEmbeddingMatrix.

This function performs k-means clustering using mini batches.

kmeans++: kmeans++ initialization. Reference : http://theory.stanford.edu/~sergei/papers/kMeansPP-soda.pdf AND http://stackoverflow.com/questions/5466323/how-exactly-does-k-means-work

random: random selection of data rows as initial centroids

References

Sculley. Web-Scale K-Means Clustering. WWW 2010, April 26–30, 2010, Raleigh, North Carolina, USA. ACM 978-1-60558-799-8/10/04.

https://github.com/mlampros/ClusterR

Author

Lampros Mouselimis and Yuwei Ni

Examples

library(SummarizedExperiment)
#> Loading required package: MatrixGenerics
#> Loading required package: matrixStats
#> 
#> Attaching package: ‘MatrixGenerics’
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> Loading required package: GenomicRanges
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, sort, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> Loading required package: S4Vectors
#> 
#> Attaching package: ‘S4Vectors’
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: GenomeInfoDb
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: ‘Biobase’
#> The following object is masked from ‘package:MatrixGenerics’:
#> 
#>     rowMedians
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     anyMissing, rowMedians
se <- SummarizedExperiment(matrix(rnorm(100), ncol=10))
mbkmeans(se, clusters = 2)
#> $centroids
#>             [,1]       [,2]       [,3]       [,4]      [,5]       [,6]
#> [1,] -0.86233366 -0.1930144  0.6247238 -0.1342908 0.0651359 -0.5407030
#> [2,] -0.09849951 -0.9753502 -0.2598674  0.1835491 0.9668127 -0.6430539
#>            [,7]        [,8]       [,9]      [,10]
#> [1,] -0.4014151  0.07637861 -0.7058427 -0.7009525
#> [2,]  1.2406141 -1.71585792  0.6492191  1.1677006
#> 
#> $WCSS_per_cluster
#> numeric(0)
#> 
#> $best_initialization
#> [1] 1
#> 
#> $iters_per_initialization
#>      [,1]
#> [1,]    2
#> 
#> $Clusters
#>  [1] 1 1 2 1 1 1 1 2 1 1
#> 
library(SingleCellExperiment)
sce <- SingleCellExperiment(matrix(rnorm(100), ncol=10))
mbkmeans(sce, clusters = 2, reduceMethod = NA)
#> $centroids
#>           [,1]       [,2]       [,3]       [,4]       [,5]        [,6]
#> [1,] 0.6842993  0.1386736 -0.4185955 0.44005471 -0.3672732 -0.06780258
#> [2,] 1.0046457 -0.6963494 -0.1473858 0.05162198  0.4315053  0.67613734
#>           [,7]       [,8]       [,9]      [,10]
#> [1,] 0.3516456  0.1282004 -0.4356472  0.1168733
#> [2,] 1.4198397 -1.4370986 -0.2415942 -0.9297087
#> 
#> $WCSS_per_cluster
#> numeric(0)
#> 
#> $best_initialization
#> [1] 1
#> 
#> $iters_per_initialization
#>      [,1]
#> [1,]    2
#> 
#> $Clusters
#>  [1] 1 1 1 2 1 1 1 1 1 2
#> 
x<-matrix(rnorm(100), ncol=10)
mbkmeans(x,clusters = 3)
#> $centroids
#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#> [1,] -0.81011435 -0.8828415  1.0545466 -0.7624309 -0.1317222  0.4530511
#> [2,] -0.07964226  0.1615282 -0.9620170  0.1830597 -0.6721411 -1.1632454
#> [3,]  0.42792790  0.6814623  0.2438877 -0.6567995  0.7017171 -0.5291350
#>             [,7]       [,8]       [,9]       [,10]
#> [1,]  1.13921503 -0.2021040 -0.2100428 -0.38370619
#> [2,] -0.08921928  0.3774196  0.6599757 -0.53591208
#> [3,] -0.78414344 -0.3207443  0.1097594  0.01639013
#> 
#> $WCSS_per_cluster
#> numeric(0)
#> 
#> $best_initialization
#> [1] 1
#> 
#> $iters_per_initialization
#>      [,1]
#> [1,]    2
#> 
#> $Clusters
#>  [1] 1 3 3 3 2 3 3 2 1 2
#>